% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_interactions.R
\name{computeInteractions}
\alias{computeInteractions}
\title{Compute H-statistic for given pairs of variables}
\usage{
computeInteractions(gbm.model, data, importance.threshold = 0,
  max.select = NULL, n.sample = 10000, interact.threshold = 0)
}
\arguments{
\item{gbm.model}{The \code{gbm.model} object used to compute the prediction}

\item{data}{the data used}

\item{importance.threshold}{the threshold from which to select the variables which interaction value will be computed, default is 0}

\item{max.select}{maximum number of variables selected for the computing of interactions}

\item{n.sample}{the number of observations used to compute H-statistic}

\item{interact.threshold}{the threshold from which to keep/retain the interactions, default is 0}
}
\value{
A dataframe which first two columns corresponds to the interacted variables and the third one corresponds to the H-statistic values for these interactions
}
\description{
Compute Friedman H-statistic for every pair of variable
}
\references{
Friedman, J. and Popescu B. (2005) \emph{Predictive Learning via Rule Ensembles} \url{http://statweb.stanford.edu/~jhf/ftp/RuleFit.pdf}
}
\seealso{
\code{\link[gbm]{interact.gbm}} for the computation of the H-statistic
}

