% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_interactions.R
\name{detectInteractions}
\alias{detectInteractions}
\title{Detect interactions}
\usage{
detectInteractions(data, max.interactions, formula.best,
  gbm.control = list(train.fraction = 0.8, family, shrinkage = 0.01,
  bag.fraction = 0.7, n.trees = 100, n.sample = 50000, depth = 5,
  importance.threshold = 0, max.select = 30),
  glm.control = list(train.fraction = 0.8, family, include = F,
  threshold.values = c(0.01, seq(0.05, 1, by = 0.05))))
}
\arguments{
\item{data}{the dataset}

\item{formula.best}{the formula of the base model provided by the user or found in \code{\link{lassoSelection}}}

\item{gbm.control}{a list containing the parameters passed to the Gradient Boosting model \itemize{
\item{n.sample}{Number of points used to compute H-statistic}
}}
}
\description{
Using Gradient Boosting Model and Friedman H-statistic to find interactions in the model.
Comparison of the performances between the base model and the interactions-augmented model using root mean squared error and gini index.
}
\references{
Friedman, J. and Popescu B. (2005) \emph{Predictive Learning via Rule Ensembles} \url{http://statweb.stanford.edu/~jhf/ftp/RuleFit.pdf}

Firedman, J. (2009) \emph{Greedy Function Approximation : A Gradient Boosting Machine} \url{https://statweb.stanford.edu/~jhf/ftp/trebst.pdf}
}
\seealso{
\code{\link[gbm]{gbm}} for GBM modeling

\code{\link[gbm]{interact.gbm}} to compute H-statitic,
}

