% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_interactions.R
\name{interactionsBestModelPlot}
\alias{interactionsBestModelPlot}
\title{Plot all the interactions included in the best augmented-model}
\usage{
interactionsBestModelPlot(x, ...)
}
\description{
The function plot  all the interactions included in the best augmented-model
}
\seealso{
\code{\link{plot.InteractionsDetection}}
}

