% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_glm.R
\name{lassoSelection}
\alias{lassoSelection}
\title{Finding the best lasso model}
\usage{
lassoSelection(data = data, target = NULL, family = "gaussian",
  type.measure = "deviance", offset = NULL, lambda.type = "lambda.min",
  seq.lambda = NULL, nlambda = 50, nfolds = 5, train.fraction = 1)
}
\arguments{
\item{data}{the dataset used. It must include the target. The offset must be removed if presents.}

\item{target}{the name of the target.}

\item{family}{distribution of the target. Available family are 'gaussian', 'gamma', 'binomial', 'multinomial', 'cox', 'mgaussian'.}

\item{type.measure}{loss to use for the cross-validation. Available values are 'mse', 'mae' (for regression), 'auc' (for classification) and 'deviance'. Default is 'deviance'.}

\item{offset}{(optionnal) : the vector of the offset values. It should be the same length as the number of rows of the dataset supplied.}

\item{lambda.type}{lambda used to use the best model. 'lambda.min' will choose lambda that optimize cross-validation performances.
'lambda.1se' gives the most regularized model such that error is within one standard error of the minimum. Default is 'lambda.min'.}

\item{seq.lambda}{(optionnal) user supplied lambda sequence}

\item{nlambda}{the number of \code{lambda} values. Default is 100.}

\item{nfolds}{The number of folds used for k-folds validation. Default is 10.}

\item{train.fraction}{fraction of data used to find the best model. Default is 1.}
}
\value{
\code{lassoSelection} returns an object of class "LassoGLM", a list consisting
\item{cvfit}{the returns object of the function \code{cv.glmnet} or \code{cv.HDtweedie} }
\item{formula}{the formula of the best model found}
\item{other}{to fill}
\item{prediction}{the predicted value by the best model}
\item{y}{true value of the target}
\item{selected_features}{vector of character of the selected features}
\item{non_selected_features}{vector of character of the non selected features}
}
\description{
Finding the best model using 5-fold cross-validation using Lasso-GLM model.
}
\examples{
data <- cbind(subset(x = mtcars, select = -mpg), y = mtcars$mpg)
lassoSelection(data, type.measure = 'mse', family = 'gaussian', lambda = 'lambda.min')
}
\seealso{
\code{\link[HDtweedie]{cv.HDtweedie}}, \code{\link[glmnet]{cv.glmnet}}
}

