% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_interactions.R
\name{perf.InteractionsDetection}
\alias{perf.InteractionsDetection}
\title{Performances of the best selected model}
\usage{
\method{perf}{InteractionsDetection}(self, criterion = "Gini",
  format = "pandoc")
}
\arguments{
\item{criterion}{Criterion used to select the best model. Values can be 'Gini' or 'RMSE'}

\item{format}{the format of the knitr table returns}
}
\value{
Table with the performances of the best model
}
\description{
Returning the performances (RMSE and Gini) of the best model selected according to criterion
}

