% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_glm.R
\name{perf.LassoGLM}
\alias{perf.LassoGLM}
\title{Give a summary of the performances of the Lasso-GLM procedure.}
\usage{
\method{perf}{LassoGLM}(self, format = "pandoc")
}
\arguments{
\item{format}{the format of the knitr table returns}
}
\value{
a table with the RMSE and the Gini index of the selected model.
}
\description{
Give a summary of the performances of the Lasso-GLM procedure.
}

