% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_interactions.R
\name{perf}
\alias{perf}
\title{Performances of the best selected model}
\usage{
perf(x, ...)
}
\description{
Returning the performances (RMSE and Gini) of the best model selected according to criterion
}
\seealso{
\code{\link[axamodel]{perf.InteractionsDetection}} for the comparison of the performances of the GLM base model and GLM
augmented model.

\code{\link[axamodel]{perf.LassoGLM}} for the performances of the GLM-LASSO selected model
}

