% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_interactions.R
\name{plot.InteractionsDetection}
\alias{plot.InteractionsDetection}
\title{Interaction plot of interactions found}
\usage{
\method{plot}{InteractionsDetection}(self, i = 1, y,
  data = eval(self$function.call$data), rev = NULL, cut.x = NULL,
  cut.trace = NULL, simplify = NULL)
}
\arguments{
\item{self}{an \code{InteractionsDetection}.}

\item{i}{the i-th best interaction according to its \emph{H-statistic} value. Default value is 1.}

\item{data}{The dataset used to perform the realize interaction plot. Default value is the dataset used in \code{detectInteractions} call.}
}
\value{
An \code{interaction.plot} of the i-th best interaction.
}
\description{
Plot the interactions found in \code{\link{detectInteractions}} procedure
}
\seealso{
\code{\link[stats]{interaction.plot}}
}

