% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_glm.R
\name{plot.LassoGLM}
\alias{plot.LassoGLM}
\title{Plotting of the best Lasso selected model}
\usage{
\method{plot}{LassoGLM}(self)
}
\arguments{
\item{self}{a LassoGLM object e.g the output of lassoSelection procedure}
}
\description{
Plots the Gini gain curve the best model selected by Lasso-GLM procedure
}

