% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_interactions.R
\name{plotBest}
\alias{plotBest}
\title{Plot Gini index of the augmented model vs base model}
\usage{
plotBest(x, ...)
}
\value{
a \code{plot_gain} plot of the gini index of both the best augmented-model and the base model
}
\description{
This method performs a plot gain (gini index) of the best augmented-model (according to \code{criterion}) vs base model
}
\details{
The function plot  all the interactions included in the best augmented-model
}
\seealso{
\code{\link[axaml]{plot_gain}} in \pkg{axaml} package
}

