% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_glm.R
\name{plotLambda}
\alias{plotLambda}
\title{plot the cross-validation curve of Lasso-GLM procedure}
\usage{
plotLambda(x, ...)
}
\arguments{
\item{self}{a LassoGLM object e.g the output of lassoSelection procedure}
}
\description{
Plotting of the performances of the model according to lambda (LASSO regularization term)
}

