% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error}
\usage{
rmse(pred, true)
}
\description{
Compute the RMSE
}
\details{
The root mean squared error is a an accuracy criterion which writes as follow:
\deqn{RMSE = \sqrt{\frac{1}{N}\sum_{i=1}^N ( y_i - \hat{y}_i)^2  } }
}

