% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_interactions.R
\name{summary.InteractionsDetection}
\alias{summary.InteractionsDetection}
\title{Return interactions of the best model}
\usage{
\method{summary}{InteractionsDetection}(self, criterion = "Gini",
  format = "pandoc")
}
\arguments{
\item{criterion}{Criterion used to select the best model. Values can be 'Gini' for gini index or 'RMSE' for Root Mean Squared Error}
}
\value{
Printing a table with the best interactions and their correspond H-statistic
}
\description{
Return all the interactions of the best model. The best model is selected according to a criterion ('Gini' or 'RMSE')
}

