% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{trainingSample}
\alias{trainingSample}
\title{Taking sample for each level of factors}
\usage{
trainingSample(data, train.fraction = 0.75)
}
\arguments{
\item{data}{The dataset}
}
\description{
Returns a training sample row names for a dataset taking draw in each level of the dataset
}
\seealso{
\code{\link[dplyr]{group_by}}
}

