% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_to_glm_functions.R
\name{cleanData}
\alias{cleanData}
\title{Cleaning the data}
\usage{
cleanData(data, n.levels = 20, perc.na = 0.2)
}
\arguments{
\item{data}{the dataset}

\item{n.levels}{the maximum number of levels allowed}

\item{per.na}{the maximum percentage of missing data allowed for a variable}
}
\value{
A clean and neat dataset
}
\description{
Removing variables with too much levels, too much NA, too few variance
}

