% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_to_glm_functions.R
\name{lassoSelection}
\alias{lassoSelection}
\title{Finding the best lasso model}
\usage{
lassoSelection(data, type.measure = "mse", family = "gaussian", offset,
  lambda = "lambda.min", train.fraction = 0.8, seq.lambda = NULL)
}
\arguments{
\item{data}{the dataset}
}
\value{
explain the return
}
\description{
this is the description
}

