% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_glm.R
\name{lassoSelection}
\alias{lassoSelection}
\title{Finding the best lasso model}
\usage{
lassoSelection(data = data, type.measure = "mse", family = "gaussian",
  offset = NULL, lambda = "lambda.min", train.fraction = 1,
  seq.lambda = NULL, nfolds = 10)
}
\arguments{
\item{type.measure}{the performance criterion use to perform cross-validation. Available values are 'mse', 'mae' (for regression), 'auc' (for classification) and deviance.}

\item{family}{distribution of the target. Available family are 'gaussian', 'gamma', 'binomial', 'multinomial', 'cox', 'mgaussian'.}

\item{offset}{name of the offset (if used)}

\item{lambda}{lambda used to use the best model. 'lambda.min' will choose lambda that optimize cross-validation performances. 'lambda.1se' will choose the lambda that}

\item{train.fraction}{fraction of data used to find the best model}

\item{seq.lambda}{user provided sequences of lambda on which to performs the cross-validation}
}
\value{
\code{lassoSelection} returns an object of class "LassoGLM", a list consisting
\item{cvfit}{the returns object of the function \code{cv.glmnet} or \code{cv.HDtweedie} }
\item{formula}{the formula of the best model found}
\item{other}{to fill}
\item{prediction}{the predicted value by the best model}
\item{y}{true value of the target}
\item{selected_features}{vector of character of the selected features}
\item{non_selected_features}{vector of character of the non selected features}
}
\description{
Finding the best model using 5-fold cross-validation using Lasso-GLM model.
}
\details{
#' @param data the dataset
}
\examples{
data = cbind(subset(x = mtcars, select = -mpg),
             y = mtcars$mpg)
lassoSelection(data, type.measure = 'mse', family = 'gaussian', lambda = 'lambda.min', train.fraction = 1)
}
\seealso{
\code{\link[HDtweedie]{cv.HDtweedie}}, \code{\link[glmnet]{cv.glmnet}}
}

