% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_interactions.R
\name{perf}
\alias{perf}
\title{Performances of the best selected model}
\usage{
perf(x, ...)
}
\arguments{
\item{criterion}{Criterion used to select the best model. Values can be 'Gini' or 'RMSE'}
}
\value{
Table with the performances of the best model
}
\description{
Returning the performances (RMSE and Gini) of the best model selected according to criterion
}

