% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_to_glm_functions.R
\name{plot.LassoGLM}
\alias{plot.LassoGLM}
\title{Plotting of the best Lasso selected model}
\usage{
\method{plot}{LassoGLM}(self)
}
\description{
Plots the Gini gain curve the best model selected by Lasso-GLM procedure
}

